Allow embedded images to be used as an icon
When using an embedded image instead of a URL, the templating library
will automatically filter it out. This makes sense as the embedded URLs
are an attack surface, but because these URLs are provided by the config
file and aren't user data it should be safe to allow them to skip
sanitization.

For examples of embedded images, I was using images generated on
https://placeholderimage.dev