Fix config not being properly loaded
The fig library didn't actually expect a direct path to a config in it's
`File` option, so if you passed a config path using the command-line
flag it wasn't actually working. I thought about adjusting the calls to
fig to make it work as expected, but the config for this app is so
simple that I don't think fig is necessary at all. Instead I just
prepopulate the struct with the defaults and then decode the yaml
directly into it. All of the validation works as it was, so this fixes
the issue and drops a dependency. Win win I suppose.